some small changes

+++ Frontend/questify-frontend/src/Components/Admin/AdminAnswers.jsx
@@ -7,7 +7,6 @@ import "../css/AdminHome.css";
 import SidebarAdmin from "./SidebarAdmin";
 
 function LastSeen({ date }) {
-  
   return (
     <div>
       <ReactTimeAgo date={date} locale="en-US" timeStyle="round" />
@@ -18,10 +17,12 @@ function AdminAnswers() {
   const [answerlists, setAnswerlists] = useState([]);
   console.log(answerlists);
 
-  const handleDelete=async(id)=>{
-   await axios.delete( `http://localhost:3000/api/admin/deleteanswer/${id}`)
-   location.reload()
-  }
+  const handleDelete = async (id) => {
+    if (window.confirm("Are you sure to delete?")) {
+      await axios.delete(`http://localhost:3000/api/admin/deleteanswer/${id}`);
+      location.reload();
+    }
+  };
 
   useEffect(() => {
     async function answer() {
@@ -78,7 +79,9 @@ function AdminAnswers() {
               </div>
               <div className="post-answer">{ReactHtmlParser(a.answer)}</div>
               <div>
-                <button className="dlt-btn" onClick={()=>handleDelete(a._id)}>Delete</button>
+                <button className="dlt-btn" onClick={() => handleDelete(a._id)}>
+                  Delete
+                </button>
               </div>
             </div>
           </>
